+++ src/components/navigation/Navigation.tsx
@@ -1,7 +1,6 @@
-import SVGIcon from '../svg-icon/SVGIcon';
+import SVGIcon from '../SvgIcon';
 import { Link, useLocation } from 'react-router-dom';
 import { NavbarIcon } from './types';
-import { cn } from '@/utils/clsx';
 import { cva } from 'class-variance-authority';
 
 const listItemVariants = cva(
