+++ src/pages/course/index.tsx
@@ -1,4 +1,4 @@
-import TabButton from '@/components/buttons/tab';
+import { TabButton } from '@/components/Button';
 import CourseListItem, { CourseListItemProps } from '@/components/pages/course/CourseListItem';
 import withAuthGuard from '@/hoc/withAuthGuard';
