+++ src/pages/culture/index.tsx
@@ -1,7 +1,6 @@
-import TabButton from '@/components/buttons/tab';
+import { TabButton } from '@/components/Button';
 import CulturePlaceItem from '@/components/pages/culture/CulturePlaceItem';
 import { usePlaceExhibition, usePlacePopup } from '@/service/place/usePlaceService';
-import { convertDateToYMD } from '@/utils';
 import clsx from 'clsx';
 import { useEffect } from 'react';
 import { useLocation, useNavigate } from 'react-router';
